feat: added core logic to sorting and matching listings to request

+++ src/main/kotlin/com/lots/multivehiclesearch/models/VehicleLot.kt
@@ -5,5 +5,5 @@ data class VehicleLot (
     val location_id: String,
     val length: Int,
     val width: Int,
-    val price_in_cents: Double
+    val price_in_cents: Long
 )
