+++ src/main/kotlin/com/lots/multivehiclesearch/models/VehicleLotResponse.kt
@@ -3,5 +3,5 @@ package com.lots.multivehiclesearch.models
 data class VehicleLotResponse (
     val location_id: String,
     val listing_ids: List<String>,
-    val total_price_in_cents: Double
+    val total_price_in_cents: Long
 )
